+++ Tests/Runtime/FlowTests/TestScript___SimpleSceneElement.cs
@@ -1,0 +1,6 @@
+namespace GameFlow.Tests
+{
+    public class TestScript___SimpleSceneElement : GameFlowElement
+    {
+    }
+}