+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Code Package for Unity
 
+## [0.0.4] - 2024-05-13
+
+- Refactor Tests Func
+- Add and Release simple GameFlowElement
+- Add Game Flow Components
+
 ## [0.0.3] - 2024-05-24
+
 - Add Game Flow Manager
 - Add Editor Generate Package
 - Add Editor Import Element
 
 ## [0.0.2] - 2024-05-15
+
 - Add Base Game Flow Runtime
 - Add Base Loading Controller
 - Create Tests for: FlowCommand & Loading Controller